+++ app/src/main/java/com/ramattec/stokemarketapp/data/remote/dto/IntraDayInfoDto.kt
@@ -1,12 +1,10 @@
 package com.ramattec.stokemarketapp.data.remote.dto
 
-import java.math.BigDecimal
-
 data class IntraDayInfoDto(
     val timestamp: String,
-    val close: BigDecimal,
-    val open: BigDecimal,
-    val low: BigDecimal,
-    val high: BigDecimal,
+    val close: Double,
+    val open: Double,
+    val low: Double,
+    val high: Double,
     val volume: Long
 )
