+++ app/src/main/java/com/ramattec/stokemarketapp/domain/model/IntraDayInfo.kt
@@ -1,9 +1,8 @@
 package com.ramattec.stokemarketapp.domain.model
 
-import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class IntraDayInfo(
     val date: LocalDateTime,
-    val close: BigDecimal
+    val close: Double
 )
